Graphical Model Visualization update
Graphical Model Visualization

We did the initial graphics model with R for stock data.

There are some further improvement needs to be done as we could
discover:

- Limit the number of stocks based on the strength of the correlation
and causation.

- Eliminate the correlation and causation with the stock itself from
the graphical model

These will be done in the next step implementation. We will do more
research and look forward to discussing with you.